+++ run-localhost.sh
@@ -1,2 +1,2 @@
 cd src-app
-python3.7 -m http.server 8000
+python3.9 -m http.server 8000