+++ src-app/change_log.md
@@ -4,6 +4,7 @@ SynDaver Symple Slicer v1.1.0
 * '''Slicing UI:''' Added ability to show overhangs when placing objects
 * '''Slicing UI:''' Added ability to enable different slicing settings under "Advanced Features"
 * '''Slicing UI:''' "Print Sequence" under "Special Modes" is now usable (no collision prevention, use with care!)
+* '''Slicing UI:''' Added ability to pause print at certain layers or insert custom G-code
 * '''Firmware:''' Added Level and LevelUp
 * '''Firmware:''' Update wireless to v1.1.1
 ## Added video stream for SynDaver Level Up