+++ src-app/change_log.md
@@ -5,10 +5,10 @@ SynDaver Symple Slicer v1.1.0
 * '''Slicing UI:''' Added ability to enable different slicing settings under "Advanced Features"
 * '''Slicing UI:''' "Print Sequence" under "Special Modes" is now usable (no collision prevention, use with care!)
 * '''Slicing UI:''' Added ability to pause print at certain layers or insert custom G-code
-* '''Firmware:''' Added Level and LevelUp
+* '''Firmware:''' Added Level and LeveL UP
 * '''Firmware:''' Update wireless to v1.1.1
-## Added video stream for SynDaver Level Up
-## Added wireless settings report for SynDaver Level Up
+## Added video stream for SynDaver LeveL UP
+## Added wireless settings report for SynDaver LeveL UP
 ## Added ability to configure default gateway and subnet mask
 ## Added ability to connect to an open access point
 ## Fixed persistent "Printer unreachable" message