+++ src-app/change_log.md
@@ -6,7 +6,7 @@ SynDaver Symple Slicer v1.1.0
 * '''Slicing UI:''' "Print Sequence" under "Special Modes" is now usable (no collision prevention, use with care!)
 * '''Slicing UI:''' Added ability to pause print at certain layers or insert custom G-code
 * '''Firmware:''' Added Level and LeveL UP
-* '''Firmware:''' Update wireless to v1.1.1
+* '''Firmware:''' Update wireless to v1.1.0
 ## Added video stream for SynDaver LeveL UP
 ## Added wireless settings report for SynDaver LeveL UP
 ## Added ability to configure default gateway and subnet mask